@@ -5,9 +5,9 @@ build-backend = "pdm.backend"
 [project]
 authors = [{ name = "Erick Friis", email = "erick@langchain.dev" }]
 license = { text = "MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 dependencies = [
-    "typer<1.0.0,>=0.9.0",
+    "typer<1.0.0,>=0.17",
     "gitpython<4,>=3",
     "langserve[all]>=0.0.51",
     "uvicorn<1.0,>=0.23",
@@ -79,6 +79,12 @@ pyupgrade.keep-runtime-typing = true
 [tool.ruff.lint.per-file-ignores]
 "tests/**" = [ "D1", "DOC", "S", "SLF",]
 "scripts/**" = [ "INP", "S",]
+
+[tool.pytest.ini_options]
+addopts = "--strict-markers --strict-config --durations=5"
+markers = [
+    "compile: mark placeholder test used to compile integration tests without running them",
+]
 
 [tool.mypy]
 plugins = ["pydantic.mypy"]
