@@ -1,8 +1,10 @@
 """Events utilities."""
 
+from __future__ import annotations
+
 import http.client
 import json
-from typing import Any, Optional, TypedDict
+from typing import Any, TypedDict
 
 import typer
 
@@ -18,10 +20,10 @@ class EventDict(TypedDict):
     """
 
     event: str
-    properties: Optional[dict[str, Any]]
+    properties: dict[str, Any] | None
 
 
-def create_events(events: list[EventDict]) -> Optional[dict[str, Any]]:
+def create_events(events: list[EventDict]) -> dict[str, Any] | None:
     """Create events.
 
     Args:
