@@ -1,12 +1,14 @@
 """Packages utilities."""
 
+from __future__ import annotations
+
 from pathlib import Path
-from typing import Any, Optional, TypedDict
+from typing import Any, TypedDict, cast
 
 from tomlkit import load
 
 
-def get_package_root(cwd: Optional[Path] = None) -> Path:
+def get_package_root(cwd: Path | None = None) -> Path:
     """Get package root directory.
 
     Args:
@@ -62,11 +64,12 @@ def get_langserve_export(filepath: Path) -> LangServeExport:
         KeyError: If the `pyproject.toml` file is missing required fields.
     """
     with filepath.open() as f:
-        data: dict[str, Any] = load(f)
+        # tomlkit types aren't amazing - treat as Dict instead
+        data = cast("dict[str, Any]", load(f))
     try:
-        module = data["tool"]["langserve"]["export_module"]
-        attr = data["tool"]["langserve"]["export_attr"]
-        package_name = data["tool"]["poetry"]["name"]
+        module = str(data["tool"]["langserve"]["export_module"])
+        attr = str(data["tool"]["langserve"]["export_attr"])
+        package_name = str(data["tool"]["poetry"]["name"])
     except KeyError as e:
         msg = "Invalid LangServe PyProject.toml"
         raise KeyError(msg) from e
