@@ -1,9 +1,10 @@
 """Script to generate migrations for the migration script."""
 
+from __future__ import annotations
+
 import json
 import pkgutil
 from pathlib import Path
-from typing import Optional
 
 import click
 
@@ -52,7 +53,7 @@ def cli() -> None:
 def generic(
     pkg1: str,
     pkg2: str,
-    output: Optional[str],
+    output: str | None,
     filter_by_all: bool,  # noqa: FBT001
     format_: str,
 ) -> None:
@@ -76,7 +77,7 @@ def generic(
     Path(output).write_text(dumped, encoding="utf-8")
 
 
-def handle_partner(pkg: str, output: Optional[str] = None) -> None:
+def handle_partner(pkg: str, output: str | None = None) -> None:
     """Handle partner package migrations."""
     migrations = get_migrations_for_partner_package(pkg)
     # Run with python 3.9+
